+++ tests/mock_hass/homeassistant/components/event/__init__.py
@@ -1,0 +1,6 @@
+class EventDeviceClass:
+    DOORBELL = "doorbell"
+    BUTTON = "button"
+
+class EventEntity:
+    pass