+++ tests/mock_hass/homeassistant/components/lock/__init__.py
@@ -1,0 +1,5 @@
+class LockEntity:
+    pass
+
+class LockEntityFeature:
+    OPEN = 1